Separate tightly coupled code in traffic cop
Short version: claw out the eyes of the traffic cop!

Long version: detecting/deciding whether it makes sense to take another
measurement does not really belong to the rest of the logic.  Proof:
This.  I actually believe that the eyes should instead be part of
victim_detector, since it may have a better way to define what constitutes
a 'good' time to take another measurement.